requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
defusedxml==0.7.1
fonttools==4.58.1
fpdf2==2.8.3
packaging==25.0
pillow==11.2.1
